Referencia/OpenDSS/IEEETestCases/13Bus/Modular/requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.1.4
plotly==5.18.0
openpyxl==3.1.2
dss-python==0.13.4
numpy==1.24.4
Flask==3.0.1